fix: DateUtil long으로 수정

+++ core/src/main/java/doro/android/core/util/DateUtil.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 object DateUtil {
 
-    fun parseMinute(time: Int): Int {
-        return (time - Date().time.toInt()) / 1000 / 60
+    fun parseMinute(time: Long): Int {
+        return (time - Date().time).toInt() / 1000 / 60
     }
 }
